Victory! css stylesheet attached
shiny.css, as you, Will, suspected is an automatic stylesheet, that is
created by default and cannot be altered. Line 9 in index.html links to
this automatic stylesheet. Lines 7 and 8 are also automatic and should
not be messed with.

you must create a custom stylesheet located somewhere in the folder www
and refer to it correspondingly, like i did in line 10 of index.html

this is a milestone point